+++ lib/acanthis.dart
@@ -9,3 +9,5 @@ export 'src/types/union.dart';
 export 'src/types/boolean.dart';
 
 export 'src/exceptions/validation_error.dart';
+export 'src/exceptions/async_exception.dart';
+export 'src/exceptions/build_validator_exception.dart';