+++ lib/src/exceptions/build_validator_exception.dart
@@ -1,0 +1,14 @@
+class BuildValidatorException implements Exception {
+
+  /// The message of the exception
+  final String message;
+
+  /// The constructor of the exception
+  BuildValidatorException(this.message);
+
+  @override
+  String toString() {
+    return 'BuildValidatorException: $message';
+  }
+
+}