+++ lib/src/types/date.dart
@@ -1,5 +1,4 @@
 import 'list.dart';
-import 'nullable.dart';
 import 'types.dart';
 import 'union.dart';
 
@@ -32,26 +31,11 @@ class AcanthisDate extends AcanthisType<DateTime> {
     return AcanthisList<DateTime>(this);
   }
 
-  /// Add a custom check to the date
-  AcanthisDate customCheck(
-      {required bool Function(DateTime value) onCheck,
-      required String error,
-      required String name}) {
-    addCheck(
-        AcanthisCheck<DateTime>(onCheck: onCheck, error: error, name: name));
-    return this;
-  }
-
   /// Add a transformation to the date to transform it using [transformation]
   AcanthisDate transform(DateTime Function(DateTime value) transformation) {
     addTransformation(
         AcanthisTransformation<DateTime>(transformation: transformation));
     return this;
-  }
-
-  /// Make the value nullable
-  AcanthisNullable<DateTime> nullable({DateTime? defaultValue}) {
-    return AcanthisNullable(this, defaultValue: defaultValue);
   }
 
   /// Create a union from the string