+++ lib/src/types/list.dart
@@ -1,4 +1,3 @@
-import 'nullable.dart';
 import 'types.dart';
 
 /// A class to validate list types
@@ -129,6 +128,8 @@ class AcanthisList<T> extends AcanthisType<List<T>> {
   }
 
   /// Add a check to the list to check if all elements are unique
+  /// 
+  /// In Zod is the same as creating a set.
   AcanthisList<T> unique() {
     addCheck(AcanthisCheck<List<T>>(
         onCheck: (toTest) => toTest.toSet().length == toTest.length,
@@ -146,24 +147,9 @@ class AcanthisList<T> extends AcanthisType<List<T>> {
     return this;
   }
 
-  /// Add a custom check to the list
-  AcanthisList<T> customCheck(
-      {required bool Function(List<T> value) onCheck,
-      required String error,
-      required String name}) {
-    addCheck(
-        AcanthisCheck<List<T>>(onCheck: onCheck, error: error, name: name));
-    return this;
-  }
-
   /// Add a transformation to the list to transform it using [transformation]
   AcanthisList<T> transform(List<T> Function(List<T> value) transformation) {
     addTransformation(AcanthisTransformation(transformation: transformation));
     return this;
-  }
-
-  /// Make the value nullable
-  AcanthisNullable<List<T>> nullable({List<T>? defaultValue}) {
-    return AcanthisNullable(this, defaultValue: defaultValue);
   }
 }