+++ test/types/map_test.dart
@@ -451,5 +451,67 @@ void main() {
       });
       expect(result.success, true);
     });
+
+    test(
+      'when creating a map validator for an object with the partial method, then only the first level should be nullable',
+      () {
+        final object = acanthis
+            .object({
+              'name': acanthis.string().min(5).max(10).encode(),
+              'attributes': acanthis.object({
+                'age': acanthis.number().gte(18),
+                'style': acanthis.object({
+                  'color': acanthis
+                      .string()
+                      .min(3)
+                      .max(10)
+              }),
+              'date': acanthis.date().min(DateTime.now())
+            }).partial()});
+
+        final result = object.tryParse({
+          'name': 'Hello',
+          'attributes': {
+            'age': null,
+            'style': {
+              'color': 'red',
+            },
+            'date': DateTime.now()
+          }
+        });
+        expect(result.success, true);
+      }
+    );
+
+    test(
+      'when creating a map validator for an object with the partial method and the deep param at true, then all the elements should be nullable',
+      () {
+        final object = acanthis
+            .object({
+              'name': acanthis.string().min(5).max(10).encode(),
+              'attributes': acanthis.object({
+                'age': acanthis.number().gte(18),
+                'style': acanthis.object({
+                  'color': acanthis
+                      .string()
+                      .min(3)
+                      .max(10)
+              }),
+              'date': acanthis.date().min(DateTime.now())
+            }).partial(deep: true)});
+
+        final result = object.tryParse({
+          'name': 'Hello',
+          'attributes': {
+            'age': null,
+            'style': {
+              'color': null,
+            },
+            'date': DateTime.now()
+          }
+        });
+        expect(result.success, true);
+      }
+    );
   });
 }